# cmd/api/docs/swagger.yaml
definitions:
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Watch Hive is an E-commerce platform to purchase Watch
  title: Go + Gin E-Commerce API Watch Hive
  version: 1.0.0
paths:
...
  /wallet:
    get:
      consumes:
      - application/json
      description: Retrieve wallet details for the authenticated user.
      parameters:
      - description: User ID
        in: header
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Wallet details retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: user_id not found or invalid user_id type
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to retrieve details
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Wallet
      tags:
      - User Wallet Management
  /wallet/history:
    get:
      consumes:
      - application/json
      description: Retrieve the transaction history for the authenticated user's wallet.
      parameters:
      - description: User ID
        in: header
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Wallet transaction history retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: user_id not found or invalid user_id type
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to retrieve transaction history
          schema:
            $ref: '#/definitions/response.Response'
      summary: Wallet History
      tags:
      - User Wallet Management
securityDefinitions:
  BearerTokenAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"